Issue #SB-15205 | Adding Utils functions
